requirements.txt
kivy==2.2.1
discord.py-self==1.9.2
aiohttp==3.8.5
certifi